requirements.txt
anthropic>=0.40.0
openpyxl>=3.1.5
PyMuPDF>=1.24.0
PyPDF2>=3.0.1
python-docx>=1.2.0
extract-msg>=0.52.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
